Fixed multiple morph targets by gloriousayu
Multiple morph targets functionality restored/fixed by gloriousayu. Number of supported bones expanded to 512 (from previous limit of 256).

+++ Source/GFur/Private/FurSkinData.h
@@ -15,8 +15,8 @@ struct FFurSkinVertex : FFurStaticVertex<TangentBasisTypeT, UVTypeT>
 		NumInfluences = bExtraBoneInfluencesT ? MAX_TOTAL_INFLUENCES : MAX_INFLUENCES_PER_STREAM,
 	};
 
-	uint8			InfluenceBones[NumInfluences];
-	uint8			InfluenceWeights[NumInfluences];
+	uint16			InfluenceBones[NumInfluences];
+	uint16			InfluenceWeights[NumInfluences];
 };
 
 /** Fur Skin Data */
